requirements.txt
fastapi==0.115.10
uvicorn==0.34.0
tensorflow==2.18.0
numpy==1.26.4
pillow==11.1.0
python-multipart==0.0.20
nest_asyncio==1.6.0
python-dotenv==1.0.1
pydantic==2.10.6
scikit-learn==1.6.1
pymongo[srv]==4.11.1